requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-simplejwt
djoser
python-dotenv